feat: merge messaging_reply into messaging_send
