+++ assistant/docs/architecture/integrations.md
@@ -23,8 +23,7 @@ graph TB
             LIST["messaging_list_conversations"]
             READ["messaging_read"]
             SEARCH["messaging_search"]
-            SEND["messaging_send"]
-            REPLY["messaging_reply"]
+            SEND["messaging_send (+ reply via thread_id)"]
             MARK_READ["messaging_mark_read"]
             ACTIVITY["messaging_analyze_activity"]
             STYLE["messaging_analyze_style"]
